stake: move actions into action module
Also delete some dead code
