+++ pd/src/components/staking.rs
@@ -4,9 +4,10 @@ use anyhow::{anyhow, Context, Result};
 use async_trait::async_trait;
 use penumbra_proto::Protobuf;
 use penumbra_stake::{
-    BaseRateData, Delegate, DelegationChanges, Epoch, IdentityKey, PendingRewardNote, RateData,
-    RewardNotes, Undelegate, Validator, ValidatorInfo, ValidatorList, ValidatorState,
-    ValidatorStatus, STAKING_TOKEN_ASSET_ID,
+    action::{Delegate, Undelegate},
+    BaseRateData, DelegationChanges, Epoch, IdentityKey, PendingRewardNote, RateData, RewardNotes,
+    Validator, ValidatorInfo, ValidatorList, ValidatorState, ValidatorStatus,
+    STAKING_TOKEN_ASSET_ID,
 };
 use penumbra_transaction::{Action, Transaction};
