+++ stake/src/action.rs
@@ -1,0 +1,9 @@
+//! Staking-related transaction actions.
+
+mod delegate;
+mod undelegate;
+mod validator_definition;
+
+pub use delegate::Delegate;
+pub use undelegate::Undelegate;
+pub use validator_definition::ValidatorDefinition;
