+++ stake/src/funding_stream.rs
@@ -65,3 +65,61 @@ impl TryFrom<pb::FundingStream> for FundingStream {
         })
     }
 }
+
+/// A set of funding streams to which validators send rewards.
+///
+/// The total commission of a validator is the sum of the individual reward rate of the
+/// [`FundingStream`]s, and cannot exceed 10000bps (100%). This property is guaranteed by the
+/// `TryFrom<Vec<FundingStream>` implementation for [`FundingStreams`], which checks the sum, and is
+/// the only way to build a non-empty [`FundingStreams`].
+#[derive(Debug, Clone, Default, Eq, PartialEq)]
+pub struct FundingStreams {
+    funding_streams: Vec<FundingStream>,
+}
+
+impl FundingStreams {
+    pub fn new() -> Self {
+        Self {
+            funding_streams: Vec::new(),
+        }
+    }
+
+    pub fn iter(&self) -> impl Iterator<Item = &FundingStream> {
+        self.funding_streams.iter()
+    }
+}
+
+impl TryFrom<Vec<FundingStream>> for FundingStreams {
+    type Error = anyhow::Error;
+
+    fn try_from(funding_streams: Vec<FundingStream>) -> Result<Self, Self::Error> {
+        if funding_streams.iter().map(|fs| fs.rate_bps).sum::<u16>() > 10_000 {
+            return Err(anyhow::anyhow!(
+                "sum of funding rates exceeds 100% (10000bps)"
+            ));
+        }
+
+        Ok(Self { funding_streams })
+    }
+}
+
+impl From<FundingStreams> for Vec<FundingStream> {
+    fn from(funding_streams: FundingStreams) -> Self {
+        funding_streams.funding_streams
+    }
+}
+
+impl AsRef<[FundingStream]> for FundingStreams {
+    fn as_ref(&self) -> &[FundingStream] {
+        &self.funding_streams
+    }
+}
+
+impl IntoIterator for FundingStreams {
+    type Item = FundingStream;
+    type IntoIter = std::vec::IntoIter<FundingStream>;
+
+    fn into_iter(self) -> Self::IntoIter {
+        self.funding_streams.into_iter()
+    }
+}
