+++ stake/src/rate.rs
@@ -1,5 +1,3 @@
-use std::collections::BTreeMap;
-
 use penumbra_proto::{
     stake::{self as pb},
     Protobuf,
@@ -7,8 +5,6 @@ use penumbra_proto::{
 use serde::{Deserialize, Serialize};
 
 use crate::{FundingStream, IdentityKey, ValidatorState};
-
-pub type RateDataById = BTreeMap<IdentityKey, RateData>;
 
 /// Describes a validator's reward rate and voting power in some epoch.
 #[derive(Debug, PartialEq, Eq, Clone, Serialize, Deserialize)]
